+++ CHANGES.rst
@@ -8,6 +8,25 @@ details, see the commit logs at https://github.com/girder/slicer_package_manager
 Next Release
 ============
 
+New Features
+------------
+
+* Support listing extension with a ``query`` parameter specifying the text expected
+  to be found in the extension name or description.
+
+Bug fixes
+---------
+
+Server
+^^^^^^
+
+* Fix creation of extension in private application.
+
+* Ensure user or administrator errors associated with API endpoints are displayed and associated
+  with HTTP error code 400 by raising a :exc:`RestException` instead of a generic :exc:`Exception`.
+
+* Update API endpoint `GET /app/{app_id}/extension` to always check user credentials.
+
 0.5.0
 =====